split gocloud reader and writer
